move most files to mdx
