+++ src/content/bar-prep.mdx
@@ -1,11 +1,12 @@
-<PublicLayout>
-
-# Bar Prep
+---
+title: Bar Prep
+slug: /bar-prep
+updatedAt: "2020-07-01"
+tags: bar-prep
+---
 
 ## Subjects
 
 * [Wills and Trusts](/bar-prep/wills-and-trusts)
 
 * [Flashcards](/bar-prep/flashcards)
-
-</PublicLayout>
