+++ src/content/delete-your-data.mdx
@@ -1,6 +1,9 @@
-<PublicLayout>
-
-# Delete Your Data
+---
+title: Delete Your Data
+slug: /delete-your-data
+updatedAt: "2020-07-01"
+tags: delete-your-data,privacy
+---
 
 > Contact us if you would like us to make Nevada Privacy Act requests on your
 > behalf.
@@ -8,5 +11,3 @@
 In 2020, both Nevada and California made it possible for you to delete your
 digital data from online companies. Our team and our affiliate attorneys can
 make these requests on your behalf.
-
-</PublicLayout>
