+++ src/content/wills-and-trusts.mdx
@@ -1,6 +1,9 @@
-<PublicLayout>
-
-# Wills and Trusts
+---
+title: Wills and Trusts
+slug: /bar-prep/wills-and-trusts
+updatedAt: "2020-07-01"
+tags: bar-prep,wills-and-trusts
+---
 
 ## A valid will
 
@@ -60,5 +63,3 @@ A will is valid if it complies to the laws where the will was executed:
 A will is made up of components that are integrated into the will that were
 present at the time of its execution. This can be proven with extrinsic
 evidence.
-
-</PublicLayout>
