+++ src/intl/en.json
@@ -8,13 +8,14 @@
         "spanish": "Spanish"
     },
     "footer": {
+        "bar_prep": "Bar Prep",
         "heading": "The Upward Mobility Law Firm",
         "text": "All content presented herein is for informational purposes only. Nothing should be construed as legal advice. Transmission and receipt of this information is not intended to create and does not constitute, an attorney-client relationship with lawyers on this platform. There is no expectation of attorney-client privilege or confidentiality of anything you may communicate to us in this forum. Do not act upon any information presented without seeking professional counsel.",
         "about": "About Us",
         "probono": "Pro Bono",
         "support": "Support",
         "practice_areas": "Practice Areas",
-        "privacy_Policy": "Privacy Policy",
+        "privacy_policy": "Privacy Policy",
         "terms": "Terms of Service",
         "switch": "Switch to ",
         "mode": "Mode",
