+++ src/pages/pro-bono/justice-for-rickie-slaughter/contact.tsx
@@ -6,7 +6,7 @@ import {
   Textarea
 } from '@chakra-ui/core';
 import { Button } from '@components/button';
-import { PublicLayout } from '@layouts/public';
+import { PublicLayout } from '@layouts/publicLayout';
 import React from 'react';
 import { useIntl } from 'gatsby-plugin-intl';
 
@@ -34,8 +34,9 @@ const ContactRickiePage = () => {
           </Text>
           <Input
             name="email"
-            placeholder={`${intl.formatMessage({ 
-              id: 'pages_pb.form.email_ph' })}`}
+            placeholder={`${intl.formatMessage({
+              id: 'pages_pb.form.email_ph'
+            })}`}
             borderColor="gray.300"
             _hover={{ borderColor: 'gray.500' }}
           />
@@ -43,8 +44,9 @@ const ContactRickiePage = () => {
             {intl.formatMessage({ id: 'pages_probono.form.message' })}
           </Text>
           <Textarea name="message"
-            placeholder={`${intl.formatMessage({ 
-              id: 'pages_pb.form.message_ph' })}`}
+            placeholder={`${intl.formatMessage({
+              id: 'pages_pb.form.message_ph'
+            })}`}
             borderColor="gray.300"
             _hover={{ borderColor: 'gray.500' }}
           />
